Split oesophageal (C15) ICD-10 codes in to SCC or AC

Using the data provided to us by CRUK, split deaths marked as due to Oesophageal cancer into deaths from the subtypes SCC and AC.

SplitOesoph(data, cruk_splits = oesoph_splits, n_deaths_var = "n_deaths")

Arguments

data

Data table containing counts of death by cause.

cruk_splits

Data table containing the percentages of oesophageal cancer that are SCC and AC.

n_deaths_var

Character - the name of the variable containing the death counts.

Value

Returns a data table that is an updated version of data with extra rows added to reflect the splitting of oesophageal cancer into two diseases.

Examples

if (FALSE) {

data <- SplitOesoph(data)

}