Lee-Carter forecast of mortality

Implements a Lee-Carter forecast separately for mortality from different causes within 10 subgroups defined by combinations of sex and IMD quintile.

MortalityForecast(
  mort_data_cause,
  folder_path = getwd(),
  cause_names = c("Chronic_obstructive_pulmonary_disease"),
  sex_focus = c("Female", "Male"),
  folder_label = NULL,
  year_range = 2001:2016,
  age_range = 11:89,
  smooth = FALSE,
  smooth_n_age = 3,
  smooth_n_year = 3,
  n_years = 50,
  standard_pop = mort.tools::esp2013,
  interpolate = TRUE
)

Arguments

mort_data_cause

Data table containing our processed cause-specific mortality rates

folder_path

where to put the folder of diagnostic files

cause_names

names of the causes to include

sex_focus

Female, Male or both

folder_label

how to identify the forecast folder

year_range

the range of years to use as the basis for forecasting

age_range

the ages for which to conduct the forecast

smooth

TRUE or FALSE

smooth_n_age

the age range for moving average - must be odd number

smooth_n_year

the year range for moving average - must be odd number

n_years

how far in future to project

standard_pop

Data table - the population structure for standardisation.

interpolate

TRUE or FALSE - should any remaining zeros in the data after smoothing be interpolated based on neighbouring values. This only works if sufficient data exists to inform the interpolation i.e. for diseases with few deaths it might have to be turned off.

Value

mort_data

Observed mortality

mort_data_smooth

Smoothed observed mortality

mx_forecast

Forecast mortality

year_plot

Age-standardised mortality trends over time

age_plot

Age pattern of mortality by year

Details

Uses the functions from the R package 'demography':

smooths the trends in mortality rates with a 2d moving average over age and year

fits the Lee-Carter model (singular value decomposition of trends)

conducts the forecast up to n_years from baseline

Examples

if (FALSE) {

mx_forecast <- MortalityForecast(
  mort_data_cause = fread("tob_death_rates_national_2019-05-06_mort.tools_1.0.0.csv")
)

}